backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["hespadas <henriqueep@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic = {extras = ["email"], version = "^2.11.5"}
pytest = "^8.3.5"
sqlalchemy = "^2.0.41"
pydantic-settings = "^2.9.1"
alembic = "^1.16.1"
psycopg2 = "^2.9.10"
pwdlib = {extras = ["argon2"], version = "^0.2.1"}
python-multipart = "^0.0.20"
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.11.11"
httpx = "^0.28.1"

[tool.ruff]
line-length = 120
exclude = ["migrations"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
